finish simplify tabnet
