+++ examples/tabnet/src/metrics.py
@@ -14,29 +14,6 @@ import torch
 
 
 def UnsupervisedLoss(y_pred, embedded_x, obf_vars, eps=1e-9):
-    """
-    Implements unsupervised loss function.
-    This differs from orginal paper as it's scaled to be batch size independent
-    and number of features reconstructed independent (by taking the mean)
-
-    Parameters
-    ----------
-    y_pred : torch.Tensor or np.array
-        Reconstructed prediction (with embeddings)
-    embedded_x : torch.Tensor
-        Original input embedded by network
-    obf_vars : torch.Tensor
-        Binary mask for obfuscated variables.
-        1 means the variable was obfuscated so reconstruction is based on this.
-    eps : float
-        A small floating point to avoid ZeroDivisionError
-        This can happen in degenerated case when a feature has only one value
-
-    Returns
-    -------
-    loss : torch float
-        Unsupervised loss, average value over batch samples.
-    """
     errors = y_pred - embedded_x
     reconstruction_errors = torch.mul(errors, obf_vars) ** 2
     batch_means = torch.mean(embedded_x, dim=0)
@@ -73,64 +50,7 @@ def UnsupervisedLossNumpy(y_pred, embedded_x, obf_vars, eps=1e-9):
 
 
 @dataclass
-class UnsupMetricContainer:
-    """Container holding a list of metrics.
-
-    Parameters
-    ----------
-    y_pred : torch.Tensor or np.array
-        Reconstructed prediction (with embeddings)
-    embedded_x : torch.Tensor
-        Original input embedded by network
-    obf_vars : torch.Tensor
-        Binary mask for obfuscated variables.
-        1 means the variables was obfuscated so
-        reconstruction is based on this.
-
-    """
-
-    metric_names: List[str]
-    prefix: str = ""
-
-    def __post_init__(self):
-        self.metrics = Metric.get_metrics_by_names(self.metric_names)
-        self.names = [self.prefix + name for name in self.metric_names]
-
-    def __call__(self, y_pred, embedded_x, obf_vars):
-        """Compute all metrics and store into a dict.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_pred : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        dict
-            Dict of metrics ({metric_name: metric_value}).
-
-        """
-        logs = {}
-        for metric in self.metrics:
-            res = metric(y_pred, embedded_x, obf_vars)
-            logs[self.prefix + metric._name] = res
-        return logs
-
-
-@dataclass
 class MetricContainer:
-    """Container holding a list of metrics.
-
-    Parameters
-    ----------
-    metric_names : list of str
-        List of metric names.
-    prefix : str
-        Prefix of metric names.
-
-    """
 
     metric_names: List[str]
     prefix: str = ""
@@ -140,21 +60,6 @@ class MetricContainer:
         self.names = [self.prefix + name for name in self.metric_names]
 
     def __call__(self, y_true, y_pred):
-        """Compute all metrics and store into a dict.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_pred : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        dict
-            Dict of metrics ({metric_name: metric_value}).
-
-        """
         logs = {}
         for metric in self.metrics:
             if isinstance(y_pred, list):
@@ -173,21 +78,6 @@ class Metric:
 
     @classmethod
     def get_metrics_by_names(cls, names):
-        """Get list of metric classes.
-
-        Parameters
-        ----------
-        cls : Metric
-            Metric class.
-        names : list
-            List of metric names.
-
-        Returns
-        -------
-        metrics : list
-            List of metric classes.
-
-        """
         available_metrics = cls.__subclasses__()
         available_names = [metric()._name for metric in available_metrics]
         metrics = []
@@ -211,21 +101,6 @@ class AUC(Metric):
         self._maximize = True
 
     def __call__(self, y_true, y_score):
-        """
-        Compute AUC of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            AUC of predictions vs targets.
-        """
         return roc_auc_score(y_true, y_score[:, 1])
 
 
@@ -239,21 +114,6 @@ class Accuracy(Metric):
         self._maximize = True
 
     def __call__(self, y_true, y_score):
-        """
-        Compute Accuracy of predictions.
-
-        Parameters
-        ----------
-        y_true: np.ndarray
-            Target matrix or vector
-        y_score: np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            Accuracy of predictions vs targets.
-        """
         y_pred = np.argmax(y_score, axis=1)
         return accuracy_score(y_true, y_pred)
 
@@ -268,21 +128,6 @@ class BalancedAccuracy(Metric):
         self._maximize = True
 
     def __call__(self, y_true, y_score):
-        """
-        Compute Accuracy of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            Accuracy of predictions vs targets.
-        """
         y_pred = np.argmax(y_score, axis=1)
         return balanced_accuracy_score(y_true, y_pred)
 
@@ -297,21 +142,6 @@ class LogLoss(Metric):
         self._maximize = False
 
     def __call__(self, y_true, y_score):
-        """
-        Compute LogLoss of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            LogLoss of predictions vs targets.
-        """
         return log_loss(y_true, y_score)
 
 
@@ -325,21 +155,6 @@ class MAE(Metric):
         self._maximize = False
 
     def __call__(self, y_true, y_score):
-        """
-        Compute MAE (Mean Absolute Error) of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            MAE of predictions vs targets.
-        """
         return mean_absolute_error(y_true, y_score)
 
 
@@ -353,21 +168,6 @@ class MSE(Metric):
         self._maximize = False
 
     def __call__(self, y_true, y_score):
-        """
-        Compute MSE (Mean Squared Error) of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            MSE of predictions vs targets.
-        """
         return mean_squared_error(y_true, y_score)
 
 
@@ -378,21 +178,6 @@ class RMSLE(Metric):
         self._maximize = False
 
     def __call__(self, y_true, y_score):
-        """
-        Compute RMSLE of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            RMSLE of predictions vs targets.
-        """
         y_score = np.clip(y_score, a_min=0, a_max=None)
         return np.sqrt(mean_squared_log_error(y_true, y_score))
 
@@ -407,24 +192,6 @@ class UnsupervisedMetric(Metric):
         self._maximize = False
 
     def __call__(self, y_pred, embedded_x, obf_vars):
-        """
-        Compute MSE (Mean Squared Error) of predictions.
-
-        Parameters
-        ----------
-        y_pred : torch.Tensor or np.array
-            Reconstructed prediction (with embeddings)
-        embedded_x : torch.Tensor
-            Original input embedded by network
-        obf_vars : torch.Tensor
-            Binary mask for obfuscated variables.
-            1 means the variables was obfuscated so reconstruction is based on this.
-
-        Returns
-        -------
-        float
-            MSE of predictions vs targets.
-        """
         loss = UnsupervisedLoss(y_pred, embedded_x, obf_vars)
         return loss.item()
 
@@ -439,24 +206,6 @@ class UnsupervisedNumpyMetric(Metric):
         self._maximize = False
 
     def __call__(self, y_pred, embedded_x, obf_vars):
-        """
-        Compute MSE (Mean Squared Error) of predictions.
-
-        Parameters
-        ----------
-        y_pred : torch.Tensor or np.array
-            Reconstructed prediction (with embeddings)
-        embedded_x : torch.Tensor
-            Original input embedded by network
-        obf_vars : torch.Tensor
-            Binary mask for obfuscated variables.
-            1 means the variables was obfuscated so reconstruction is based on this.
-
-        Returns
-        -------
-        float
-            MSE of predictions vs targets.
-        """
         return UnsupervisedLossNumpy(
             y_pred,
             embedded_x,
@@ -474,38 +223,10 @@ class RMSE(Metric):
         self._maximize = False
 
     def __call__(self, y_true, y_score):
-        """
-        Compute RMSE (Root Mean Squared Error) of predictions.
-
-        Parameters
-        ----------
-        y_true : np.ndarray
-            Target matrix or vector
-        y_score : np.ndarray
-            Score matrix or vector
-
-        Returns
-        -------
-        float
-            RMSE of predictions vs targets.
-        """
         return np.sqrt(mean_squared_error(y_true, y_score))
 
 
 def check_metrics(metrics):
-    """Check if custom metrics are provided.
-
-    Parameters
-    ----------
-    metrics : list of str or classes
-        List with built-in metrics (str) or custom metrics (classes).
-
-    Returns
-    -------
-    val_metrics : list of str
-        List of metric names.
-
-    """
     val_metrics = []
     for metric in metrics:
         if isinstance(metric, str):
