+++ examples/tabnet/src/tab_model.py
@@ -7,11 +7,11 @@ from utils import (
     validate_eval_set,
     create_dataloaders,
     define_device,
-    ComplexEncoder,
-    check_input,
-    check_warm_start,
     create_group_matrix,
-    check_embedding_parameters
+    check_embedding_parameters,
+    infer_output_dim,
+    infer_multitask_output,
+    check_output_dim
 )
 from callbacks import (
     CallbackContainer,
@@ -20,7 +20,6 @@ from callbacks import (
     LRSchedulerCallback,
 )
 from metrics import MetricContainer, check_metrics
-from multiclass_utils import infer_output_dim, infer_multitask_output, check_output_dim
 
 from scipy.special import softmax
 from torch.utils.data import DataLoader
@@ -31,15 +30,9 @@ import torch
 from torch.nn.utils import clip_grad_norm_
 import numpy as np
 from scipy.sparse import csc_matrix
-from abc import abstractmethod
 
 from sklearn.base import BaseEstimator
 
-import io
-import json
-from pathlib import Path
-import shutil
-import zipfile
 import warnings
 import copy
 
@@ -94,36 +87,6 @@ class TabModel(BaseEstimator):
                                                     self.cat_emb_dim)
         self.cat_dims, self.cat_idxs, self.cat_emb_dim = updated_params
 
-    def __update__(self, **kwargs):
-        """
-        Updates parameters.
-        If does not already exists, creates it.
-        Otherwise overwrite with warnings.
-        """
-        update_list = [
-            "cat_dims",
-            "cat_emb_dim",
-            "cat_idxs",
-            "input_dim",
-            "mask_type",
-            "n_a",
-            "n_d",
-            "n_independent",
-            "n_shared",
-            "n_steps",
-            "grouped_features",
-        ]
-        for var_name, value in kwargs.items():
-            if var_name in update_list:
-                try:
-                    exec(f"global previous_val; previous_val = self.{var_name}")
-                    if previous_val != value:  # noqa
-                        wrn_msg = f"Pretraining: {var_name} changed from {previous_val} to {value}"  # noqa
-                        warnings.warn(wrn_msg)
-                        exec(f"self.{var_name} = value")
-                except AttributeError:
-                    exec(f"self.{var_name} = value")
-
     def fit(
         self,
         X_train,
@@ -141,58 +104,10 @@ class TabModel(BaseEstimator):
         drop_last=True,
         callbacks=None,
         pin_memory=True,
-        from_unsupervised=None,
         warm_start=False,
         augmentations=None,
         compute_importance=True
     ):
-        """Train a neural network stored in self.network
-        Using train_dataloader for training data and
-        valid_dataloader for validation.
-
-        Parameters
-        ----------
-        X_train : np.ndarray
-            Train set
-        y_train : np.array
-            Train targets
-        eval_set : list of tuple
-            List of eval tuple set (X, y).
-            The last one is used for early stopping
-        eval_name : list of str
-            List of eval set names.
-        eval_metric : list of str
-            List of evaluation metrics.
-            The last metric is used for early stopping.
-        loss_fn : callable or None
-            a PyTorch loss function
-        weights : bool or dictionnary
-            0 for no balancing
-            1 for automated balancing
-            dict for custom weights per class
-        max_epochs : int
-            Maximum number of epochs during training
-        patience : int
-            Number of consecutive non improving epoch before early stopping
-        batch_size : int
-            Training batch size
-        virtual_batch_size : int
-            Batch size for Ghost Batch Normalization (virtual_batch_size < batch_size)
-        num_workers : int
-            Number of workers used in torch.utils.data.DataLoader
-        drop_last : bool
-            Whether to drop last batch during training
-        callbacks : list of callback function
-            List of custom callbacks
-        pin_memory: bool
-            Whether to set pin_memory to True or False during training
-        from_unsupervised: unsupervised trained model
-            Use a previously self supervised model as starting weights
-        warm_start: bool
-            If True, current model parameters are used to start training
-        compute_importance : bool
-            Whether to compute feature importance
-        """
         # update model name
         self.max_epochs = max_epochs
         self.patience = patience
@@ -217,9 +132,6 @@ class TabModel(BaseEstimator):
         else:
             self.loss_fn = loss_fn
 
-        check_input(X_train)
-        check_warm_start(warm_start, from_unsupervised)
-
         self.update_fit_params(
             X_train,
             y_train,
@@ -234,21 +146,16 @@ class TabModel(BaseEstimator):
             X_train, y_train, eval_set
         )
 
-        if from_unsupervised is not None:
-            # Update parameters to match self pretraining
-            self.__update__(**from_unsupervised.get_params())
-
         if not hasattr(self, "network") or not warm_start:
             # model has never been fitted before of warm_start is False
             self._set_network()
-        self._update_network_params()
+        self.network.virtual_batch_size = self.virtual_batch_size
         self._set_metrics(eval_metric, eval_names)
-        self._set_optimizer()
+        self._optimizer = self.optimizer_fn(
+            self.network.parameters(), **self.optimizer_params
+        )
         self._set_callbacks(callbacks)
 
-        if from_unsupervised is not None:
-            self.load_weights_from_unsupervised(from_unsupervised)
-            warnings.warn("Loading weights from unsupervised pretraining")
         # Call method on_train_begin for all callbacks
         self._callback_container.on_train_begin()
 
@@ -281,19 +188,6 @@ class TabModel(BaseEstimator):
             self.feature_importances_ = self._compute_feature_importances(X_train)
 
     def predict(self, X):
-        """
-        Make predictions on a batch (valid)
-
-        Parameters
-        ----------
-        X : a :tensor: `torch.Tensor` or matrix: `scipy.sparse.csr_matrix`
-            Input data
-
-        Returns
-        -------
-        predictions : np.array
-            Predictions of the regression problem
-        """
         self.network.eval()
 
         if scipy.sparse.issparse(X):
@@ -319,23 +213,6 @@ class TabModel(BaseEstimator):
         return self.predict_func(res)
 
     def explain(self, X, normalize=False):
-        """
-        Return local explanation
-
-        Parameters
-        ----------
-        X : tensor: `torch.Tensor` or matrix: `scipy.sparse.csr_matrix`
-            Input data
-        normalize : bool (default False)
-            Wheter to normalize so that sum of features are equal to 1
-
-        Returns
-        -------
-        M_explain : matrix
-            Importance per sample, per columns.
-        masks : matrix
-            Sparse matrix showing attention masks used by network.
-        """
         self.network.eval()
 
         if scipy.sparse.issparse(X):
@@ -396,94 +273,7 @@ class TabModel(BaseEstimator):
         for attr_name, attr_value in class_attrs.items():
             setattr(self, attr_name, attr_value)
 
-    def save_model(self, path):
-        """Saving TabNet model in two distinct files.
-
-        Parameters
-        ----------
-        path : str
-            Path of the model.
-
-        Returns
-        -------
-        str
-            input filepath with ".zip" appended
-
-        """
-        saved_params = {}
-        init_params = {}
-        for key, val in self.get_params().items():
-            if isinstance(val, type):
-                # Don't save torch specific params
-                continue
-            else:
-                init_params[key] = val
-        saved_params["init_params"] = init_params
-
-        class_attrs = {
-            "preds_mapper": self.preds_mapper
-        }
-        saved_params["class_attrs"] = class_attrs
-
-        # Create folder
-        Path(path).mkdir(parents=True, exist_ok=True)
-
-        # Save models params
-        with open(Path(path).joinpath("model_params.json"), "w", encoding="utf8") as f:
-            json.dump(saved_params, f, cls=ComplexEncoder)
-
-        # Save state_dict
-        torch.save(self.network.state_dict(), Path(path).joinpath("network.pt"))
-        shutil.make_archive(path, "zip", path)
-        shutil.rmtree(path)
-        print(f"Successfully saved model at {path}.zip")
-        return f"{path}.zip"
-
-    def load_model(self, filepath):
-        """Load TabNet model.
-
-        Parameters
-        ----------
-        filepath : str
-            Path of the model.
-        """
-        try:
-            with zipfile.ZipFile(filepath) as z:
-                with z.open("model_params.json") as f:
-                    loaded_params = json.load(f)
-                    loaded_params["init_params"]["device_name"] = self.device_name
-                with z.open("network.pt") as f:
-                    try:
-                        saved_state_dict = torch.load(f, map_location=self.device)
-                    except io.UnsupportedOperation:
-                        # In Python <3.7, the returned file object is not seekable (which at least
-                        # some versions of PyTorch require) - so we'll try buffering it in to a
-                        # BytesIO instead:
-                        saved_state_dict = torch.load(
-                            io.BytesIO(f.read()),
-                            map_location=self.device,
-                        )
-        except KeyError:
-            raise KeyError("Your zip file is missing at least one component")
-
-        self.__init__(**loaded_params["init_params"])
-
-        self._set_network()
-        self.network.load_state_dict(saved_state_dict)
-        self.network.eval()
-        self.load_class_attrs(loaded_params["class_attrs"])
-
-        return
-
     def _train_epoch(self, train_loader):
-        """
-        Trains one epoch of the network in self.network
-
-        Parameters
-        ----------
-        train_loader : a :class: `torch.utils.data.Dataloader`
-            DataLoader with train set
-        """
         self.network.train()
 
         for batch_idx, (X, y) in enumerate(train_loader):
@@ -499,23 +289,6 @@ class TabModel(BaseEstimator):
         return
 
     def _train_batch(self, X, y):
-        """
-        Trains one batch of data
-
-        Parameters
-        ----------
-        X : torch.Tensor
-            Train matrix
-        y : torch.Tensor
-            Target matrix
-
-        Returns
-        -------
-        batch_outs : dict
-            Dictionnary with "y": target and "score": prediction scores.
-        batch_logs : dict
-            Dictionnary with "batch_size" and "loss".
-        """
         batch_logs = {"batch_size": X.shape[0]}
 
         X = X.to(self.device).float()
@@ -544,16 +317,6 @@ class TabModel(BaseEstimator):
         return batch_logs
 
     def _predict_epoch(self, name, loader):
-        """
-        Predict an epoch and update metrics.
-
-        Parameters
-        ----------
-        name : str
-            Name of the validation set
-        loader : torch.utils.data.Dataloader
-                DataLoader with validation set
-        """
         # Setting network on evaluation mode
         self.network.eval()
 
@@ -574,19 +337,6 @@ class TabModel(BaseEstimator):
         return
 
     def _predict_batch(self, X):
-        """
-        Predict one batch of data.
-
-        Parameters
-        ----------
-        X : torch.Tensor
-            Owned products
-
-        Returns
-        -------
-        np.array
-            model scores
-        """
         X = X.to(self.device).float()
 
         # compute model output
@@ -632,16 +382,6 @@ class TabModel(BaseEstimator):
         )
 
     def _set_metrics(self, metrics, eval_names):
-        """Set attributes relative to the metrics.
-
-        Parameters
-        ----------
-        metrics : list of str
-            List of eval metric names.
-        eval_names : list of str
-            List of eval set names.
-
-        """
         metrics = metrics or [self._default_metric]
 
         metrics = check_metrics(metrics)
@@ -664,14 +404,6 @@ class TabModel(BaseEstimator):
         )
 
     def _set_callbacks(self, custom_callbacks):
-        """Setup the callbacks functions.
-
-        Parameters
-        ----------
-        custom_callbacks : list of func
-            List of callback functions.
-
-        """
         # Setup default callbacks history, early stopping and scheduler
         callbacks = []
         self.history = History(self, verbose=self.verbose)
@@ -706,32 +438,7 @@ class TabModel(BaseEstimator):
         self._callback_container = CallbackContainer(callbacks)
         self._callback_container.set_trainer(self)
 
-    def _set_optimizer(self):
-        """Setup optimizer."""
-        self._optimizer = self.optimizer_fn(
-            self.network.parameters(), **self.optimizer_params
-        )
-
     def _construct_loaders(self, X_train, y_train, eval_set):
-        """Generate dataloaders for train and eval set.
-
-        Parameters
-        ----------
-        X_train : np.array
-            Train set.
-        y_train : np.array
-            Train targets.
-        eval_set : list of tuple
-            List of eval tuple set (X, y).
-
-        Returns
-        -------
-        train_dataloader : `torch.utils.data.Dataloader`
-            Training dataloader.
-        valid_dataloaders : list of `torch.utils.data.Dataloader`
-            List of validation dataloaders.
-
-        """
         # all weights are not allowed for this type of model
         y_train_mapped = self.prepare_target(y_train)
         for i, (X, y) in enumerate(eval_set):
@@ -751,82 +458,10 @@ class TabModel(BaseEstimator):
         return train_dataloader, valid_dataloaders
 
     def _compute_feature_importances(self, X):
-        """Compute global feature importance.
-
-        Parameters
-        ----------
-        loader : `torch.utils.data.Dataloader`
-            Pytorch dataloader.
-
-        """
         M_explain, _ = self.explain(X, normalize=False)
         sum_explain = M_explain.sum(axis=0)
         feature_importances_ = sum_explain / np.sum(sum_explain)
         return feature_importances_
-
-    def _update_network_params(self):
-        self.network.virtual_batch_size = self.virtual_batch_size
-
-    @abstractmethod
-    def update_fit_params(self, X_train, y_train, eval_set, weights):
-        """
-        Set attributes relative to fit function.
-
-        Parameters
-        ----------
-        X_train : np.ndarray
-            Train set
-        y_train : np.array
-            Train targets
-        eval_set : list of tuple
-            List of eval tuple set (X, y).
-        weights : bool or dictionnary
-            0 for no balancing
-            1 for automated balancing
-        """
-        raise NotImplementedError(
-            "users must define update_fit_params to use this base class"
-        )
-
-    @abstractmethod
-    def compute_loss(self, y_score, y_true):
-        """
-        Compute the loss.
-
-        Parameters
-        ----------
-        y_score : a :tensor: `torch.Tensor`
-            Score matrix
-        y_true : a :tensor: `torch.Tensor`
-            Target matrix
-
-        Returns
-        -------
-        float
-            Loss value
-        """
-        raise NotImplementedError(
-            "users must define compute_loss to use this base class"
-        )
-
-    @abstractmethod
-    def prepare_target(self, y):
-        """
-        Prepare target before training.
-
-        Parameters
-        ----------
-        y : a :tensor: `torch.Tensor`
-            Target matrix.
-
-        Returns
-        -------
-        `torch.Tensor`
-            Converted target matrix.
-        """
-        raise NotImplementedError(
-            "users must define prepare_target to use this base class"
-        )
 
 
 class TabNetClassifier(TabModel):
@@ -837,20 +472,6 @@ class TabNetClassifier(TabModel):
         self._default_metric = 'accuracy'
 
     def weight_updater(self, weights):
-        """
-        Updates weights dictionary according to target_mapper.
-
-        Parameters
-        ----------
-        weights : bool or dict
-            Given weights for balancing training.
-
-        Returns
-        -------
-        bool or dict
-            Same bool if weights are bool, updated dict otherwise.
-
-        """
         if isinstance(weights, int):
             return weights
         elif isinstance(weights, dict):
@@ -896,19 +517,6 @@ class TabNetClassifier(TabModel):
         return np.vectorize(self.preds_mapper.get)(outputs.astype(str))
 
     def predict_proba(self, X):
-        """
-        Make predictions for classification on a batch (valid)
-
-        Parameters
-        ----------
-        X : a :tensor: `torch.Tensor` or matrix: `scipy.sparse.csr_matrix`
-            Input data
-
-        Returns
-        -------
-        res : np.ndarray
-
-        """
         self.network.eval()
 
         if scipy.sparse.issparse(X):
@@ -991,22 +599,6 @@ class TabNetMultiTaskClassifier(TabModel):
         return y_mapped
 
     def compute_loss(self, y_pred, y_true):
-        """
-        Computes the loss according to network output and targets
-
-        Parameters
-        ----------
-        y_pred : list of tensors
-            Output of network
-        y_true : LongTensor
-            Targets label encoded
-
-        Returns
-        -------
-        loss : torch.Tensor
-            output of loss function(s)
-
-        """
         loss = 0
         y_true = y_true.long()
         if isinstance(self.loss_fn, list):
@@ -1051,19 +643,6 @@ class TabNetMultiTaskClassifier(TabModel):
         filter_weights(self.updated_weights)
 
     def predict(self, X):
-        """
-        Make predictions on a batch (valid)
-
-        Parameters
-        ----------
-        X : a :tensor: `torch.Tensor` or matrix: `scipy.sparse.csr_matrix`
-            Input data
-
-        Returns
-        -------
-        results : np.array
-            Predictions of the most probable class
-        """
         self.network.eval()
 
         if scipy.sparse.issparse(X):
@@ -1104,19 +683,6 @@ class TabNetMultiTaskClassifier(TabModel):
         return results
 
     def predict_proba(self, X):
-        """
-        Make predictions for classification on a batch (valid)
-
-        Parameters
-        ----------
-        X : a :tensor: `torch.Tensor` or matrix: `scipy.sparse.csr_matrix`
-            Input data
-
-        Returns
-        -------
-        res : list of np.ndarray
-
-        """
         self.network.eval()
 
         if scipy.sparse.issparse(X):
